storage: Move updates logic to its own file
